% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{voro3d-package}
\alias{voro3d}
\alias{voro3d-package}
\title{Voronoi Diagrams in 3D}
\description{
Create voronoi diagrams in three dimensions from point data.
}
\author{
\strong{Maintainer}: Basil Eric Rabi \email{basil.rabi@tmc.nickelasia.com}

}
