% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{voronoi}
\alias{voronoi}
\title{Create Voronoi Diagram}
\usage{
voronoi(x, y, z, containerRatio)
}
\arguments{
\item{x}{numeric vector of the x-coordinates of the points}

\item{y}{numeric vector of the y-coordinates of the points}

\item{z}{numeric vector of the z-coordinates of the points}

\item{containerRatio}{numeric ratio between the length of the container to
be created and the length of the bounding box of the points}
}
\value{
character vector defining the voronoi cells (polyhedral surface)
  in well-known text
}
\description{
Create cell-based voronoi diagram using three-dimensional points. The
  polyhedral surface of each cell is defined in well-known text format.
}
